refactor(accounts): remove remaining delegate module references
