+++ backend/main/lib/groupher_server/accounts/mailboxes.ex
@@ -1,24 +1,45 @@
 defmodule GroupherServer.Accounts.Mailboxes do
   @moduledoc false
+  import Ecto.Query, warn: false
 
-  alias Helper.Types, as: T
+  import Helper.Utils, only: [done: 1]
 
-  alias GroupherServer.Accounts.Delegate.Mailbox
-  alias GroupherServer.Accounts.Model.User
+  alias GroupherServer.Delivery
 
-  @spec mailbox_status(User.t()) :: T.domain_res(map())
-  def mailbox_status(user), do: Mailbox.mailbox_status(user)
+  alias GroupherServer.Accounts.Model.{Embeds, User}
+  alias Helper.ORM
 
-  @spec update_mailbox_status(T.id()) :: T.domain_res(User.t())
-  def update_mailbox_status(user_id), do: Mailbox.update_mailbox_status(user_id)
+  @default_mailbox_status Embeds.UserMailbox.default_status()
 
-  @spec mark_read(atom(), [T.id()], User.t()) :: T.domain_res(map())
-  def mark_read(type, ids, user), do: Mailbox.mark_read(type, ids, user)
+  def mailbox_status(%User{mailbox: nil}), do: @default_mailbox_status |> done
+  def mailbox_status(%User{mailbox: mailbox}), do: mailbox |> done
 
-  @spec mark_read_all(atom(), User.t()) :: T.domain_res(map())
-  def mark_read_all(type, user), do: Mailbox.mark_read_all(type, user)
+  def mark_read(type, ids, %User{} = user) do
+    Delivery.mark_read(type, ids, user)
+  end
 
-  @spec paged_mailbox_messages(atom(), User.t(), map()) :: T.domain_res(T.paged_data())
-  def paged_mailbox_messages(type, user, filter),
-    do: Mailbox.paged_mailbox_messages(type, user, filter)
+  def mark_read_all(type, %User{} = user), do: Delivery.mark_read_all(type, user)
+
+  def paged_mailbox_messages(type, user, filter) do
+    Delivery.fetch(type, user, filter)
+  end
+
+  @doc "update messages count in mailbox"
+  def update_mailbox_status(user_id) do
+    with {:ok, user} <- ORM.find(User, user_id),
+         {:ok, unread_mentions_count} <- Delivery.unread_count(:mention, user_id),
+         {:ok, unread_notifications_count} <- Delivery.unread_count(:notification, user_id) do
+      unread_total_count = unread_mentions_count + unread_notifications_count
+      is_empty = unread_total_count < 1
+
+      mailbox = %{
+        unread_mentions_count: unread_mentions_count,
+        unread_notifications_count: unread_notifications_count,
+        unread_total_count: unread_total_count,
+        is_empty: is_empty
+      }
+
+      user |> ORM.update_embed(:mailbox, mailbox)
+    end
+  end
 end
