+++ backend/main/lib/groupher_server/accounts/publishes.ex
@@ -1,24 +1,40 @@
 defmodule GroupherServer.Accounts.Publishes do
-  @moduledoc false
+  @moduledoc """
+  user followers / following related
+  """
+  import Ecto.Query, warn: false
+  import Helper.Utils, only: [plural: 1]
 
-  alias Helper.Types, as: T
-
-  alias GroupherServer.Accounts.Delegate.Publish
   alias GroupherServer.Accounts.Model.User
+  alias GroupherServer.CMS
 
-  @spec paged_published_articles(User.t(), T.article_thread(), map()) ::
-          T.domain_res(T.paged_data())
-  def paged_published_articles(user, thread, filter),
-    do: Publish.paged_published_articles(user, thread, filter)
+  alias Helper.ORM
 
-  @spec paged_published_comments(User.t(), T.article_thread(), map()) ::
-          T.domain_res(T.paged_data())
-  def paged_published_comments(user, thread, filter),
-    do: Publish.paged_published_comments(user, thread, filter)
+  @doc """
+  get paged published contents of a user
+  """
+  def paged_published_articles(%User{id: user_id}, thread, filter) do
+    with {:ok, user} <- ORM.find(User, user_id) do
+      CMS.Articles.paged_published(thread, filter, user)
+    end
+  end
 
-  @spec paged_published_comments(User.t(), T.article_thread()) :: T.domain_res(T.paged_data())
-  def paged_published_comments(user, thread), do: Publish.paged_published_comments(user, thread)
+  @doc """
+  update published articles count in user meta
+  """
+  def update_published_states(%User{} = user, thread) do
+    filter = %{page: 1, size: 1}
 
-  @spec update_published_states(User.t(), T.article_thread()) :: T.domain_res(User.t())
-  def update_published_states(user, thread), do: Publish.update_published_states(user, thread)
+    with {:ok, paged_articles} <- CMS.Articles.paged_published(thread, filter, user) do
+      ORM.update_meta(user, %{:"published_#{plural(thread)}_count" => paged_articles.total_count})
+    end
+  end
+
+  def paged_published_comments(user, filter) do
+    CMS.Comments.paged_published_comments(user, filter)
+  end
+
+  def paged_published_comments(user, thread, filter) do
+    CMS.Comments.paged_published_comments(user, thread, filter)
+  end
 end
