+++ backend/main/lib/groupher_server/accounts/searches.ex
@@ -1,10 +1,23 @@
 defmodule GroupherServer.Accounts.Searches do
-  @moduledoc false
+  @moduledoc """
+  search for users
+  """
 
-  alias Helper.Types, as: T
+  import Helper.Utils, only: [done: 1]
+  import Ecto.Query, warn: false
 
-  alias GroupherServer.Accounts.Delegate.Search
+  alias GroupherServer.Accounts.Model.User
+  alias Helper.ORM
 
-  @spec search_users(map()) :: T.domain_res(T.paged_users())
-  def search_users(args), do: Search.search_users(args)
+  @search_items_count 15
+
+  @doc """
+  search community by title
+  """
+  def search_users(%{name: name} = _args) do
+    User
+    |> where([c], ilike(c.nickname, ^"%#{name}%"))
+    |> ORM.paginator(page: 1, size: @search_items_count)
+    |> done()
+  end
 end
